streamlit-app/requirements.txt
streamlit
pandas
numpy
gspread==6.1.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
statsmodels
